Move getCaseValues.
